[webui] Move ProjectStatusCalculator to app/models/
Split the file app/jobs/project_status_calculator.rb in two files:
- app/models/project_status/calculator.rb
- app/models/project_status/pack_info.rb

Mobprogrammed by: @mdeniz, @DavidKang and @eduardoj
